+++ app-src/LoomSE/Models/Scene.js
