--- app-src/LoomSE/model/index.js
@@ -1,2 +1,0 @@
-export { Story } from './Story';
-export { Scene } from './Scene';
